store: runs, events, and active-context repository
- Runs: create (records run.started as event 1 with provider, model, and
  system prompt), get, list, previous run, running run(s), complete and
  fail. Terminal events are written in the same transaction as the status
  change; only running runs change state; one running run per project
  (explicit RunInProgress error).
- Events: append with store-assigned sequences (1, 2, 3, ... per run),
  only while the run is running; read all or after a sequence cursor.
  Stored payloads are re-validated into typed EventData on read.
- Active context: replace_context writes a running run's whole membership
  atomically, ordered by node creation (created_at, id) with its source
  (carried/jev); members must be unique nodes of the run's project. Leaving
  a node out removes only its membership row, never the node. A finished
  run's context is frozen history.
- Integration tests cover identity retention, per-run monotonic
  sequences, lifecycle freezing, context ordering/isolation/removal, and
  restart persistence including crashed-run discovery.

Closes #6

+++ crates/dagos-core/src/domain/event.rs
@@ -15,7 +15,7 @@ use super::time::Timestamp;
 pub struct Event {
     pub id: EventId,
     pub run_id: RunId,
-    pub sequence: u64,
+    pub sequence: u32,
     #[serde(flatten)]
     pub data: EventData,
     pub created_at: Timestamp,
